提交

+++ classrepair-time-table/src/main/java/team/weacsoft/timetable/entity/response/ReceiverMailList.java
@@ -24,7 +24,7 @@ public class ReceiverMailList {
     /**
      * 工作人员学号/工号
      */
-    private int receiverIdentityId=0;
+    private long receiverIdentityId=0;
 
     /**
      * 工作人员电话
